// idcs_portal/src/store/models/consoleUI/consoleUI.ts
// INTEL CONFIDENTIAL
// Copyright (C) 2023 Intel Corporation

export default interface ConsoleUI {
  id: string
  familyId: string
  vendorId: string
  name: string
  description: string
  category: string
  service: string
  familyDisplayName: string
  instanceType: string
  displayName: string
  url: string
}